added gates that open on all enemy kill, impl enemies remaining text, impl every enemy actor used for battle hall
i also added some proper symbols for a couple actors and utils, not necessary but a nice to have.
many actors used in battle quest had to be slightly modified in order to work within the context of the hall, some of which re-use logic applied to enemy rando.
the gates had to be implemented in a bit of a weird way, as i dont think its possible to re-use the scene clear flag within the same room and scene. so instead, i maintain (to the best capabilities) a list of actors spawned by the battle hall system.